app/requirements.txt
flask
flask-cors
mysql-connector-python
requests
python-dotenv
apscheduler
